TableNG: Frames field range support + SparklineCell min/max
