+++ packages/grafana-data/src/field/fieldOverrides.test.ts
@@ -20,7 +20,6 @@ import {
   applyFieldOverrides,
   applyRawFieldOverrides,
   FieldOverrideEnv,
-  findNumericFieldMinMax,
   getLinksSupplier,
   setDynamicConfigValue,
   setFieldConfigDefaults,
@@ -79,68 +78,6 @@ locationUtil.initialize({
   config: { appSubUrl: '/subUrl' } as GrafanaConfig,
   getVariablesUrlParams: jest.fn(),
   getTimeRangeForUrl: jest.fn(),
-});
-
-describe('Global MinMax', () => {
-  it('find global min max', () => {
-    const f0 = new ArrayDataFrame<{ title: string; value: number; value2: number | null }>([
-      { title: 'AAA', value: 100, value2: 1234 },
-      { title: 'BBB', value: -20, value2: null },
-      { title: 'CCC', value: 200, value2: 1000 },
-    ]);
-
-    const minmax = findNumericFieldMinMax([f0]);
-    expect(minmax.min).toEqual(-20);
-    expect(minmax.max).toEqual(1234);
-  });
-
-  it('find global min max when all values are zero', () => {
-    const f0 = new ArrayDataFrame<{ title: string; value: number; value2: number | null }>([
-      { title: 'AAA', value: 0, value2: 0 },
-      { title: 'CCC', value: 0, value2: 0 },
-    ]);
-
-    const minmax = findNumericFieldMinMax([f0]);
-    expect(minmax.min).toEqual(0);
-    expect(minmax.max).toEqual(0);
-  });
-
-  describe('when value is null', () => {
-    it('then global min max should be null', () => {
-      const frame = toDataFrame({
-        fields: [
-          { name: 'Time', type: FieldType.time, values: [1] },
-          { name: 'Value', type: FieldType.number, values: [null] },
-        ],
-      });
-      const { min, max } = findNumericFieldMinMax([frame]);
-
-      expect(min).toBe(null);
-      expect(max).toBe(null);
-    });
-  });
-
-  describe('when values are zero', () => {
-    it('then global min max should be correct', () => {
-      const frame = toDataFrame({
-        fields: [
-          { name: 'Time', type: FieldType.time, values: [1, 2] },
-          { name: 'Value', type: FieldType.number, values: [1, 2] },
-        ],
-      });
-      const frame2 = toDataFrame({
-        fields: [
-          { name: 'Time', type: FieldType.time, values: [1, 2] },
-          { name: 'Value', type: FieldType.number, values: [0, 0] },
-        ],
-      });
-
-      const { min, max } = findNumericFieldMinMax([frame, frame2]);
-
-      expect(min).toBe(0);
-      expect(max).toBe(2);
-    });
-  });
 });
 
 describe('applyFieldOverrides', () => {
