+++ packages/grafana-data/src/internal/index.ts
@@ -100,7 +100,7 @@ export { type PluginAddedLinksConfigureFunc, type PluginExtensionEventHelpers } 
 export { getStreamingFrameOptions } from '../dataframe/StreamingDataFrame';
 export { fieldIndexComparer } from '../field/fieldComparers';
 export { decoupleHideFromState } from '../field/fieldState';
-export { findNumericFieldMinMax } from '../field/fieldOverrides';
+export { findNumericFieldMinMax } from '../field/scale';
 export { type PanelOptionsSupplier } from '../panel/PanelPlugin';
 export { sanitize, sanitizeUrl } from '../text/sanitize';
 export { type NestedValueAccess, type NestedPanelOptions, isNestedPanelOptions } from '../utils/OptionsUIBuilders';
